third_party/blink/web_tests/editing/execCommand/indent/indent_empty_quote.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script>
<div id="log"></div>
<script>
test(() => {
  assert_selection(
    '<div contenteditable><blockquote>|<br></blockquote></div>',
    'indent',
    '<div contenteditable><blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;"><blockquote>|<br></blockquote></blockquote></div>');
}, 'indent empty <blockquote>');
test(() => {
  assert_selection(
    '<div contenteditable><blockquote>1|</blockquote></div>',
    'indent',
    '<div contenteditable><blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;"><blockquote>1|</blockquote></blockquote></div>');
}, 'indent <blockquote> with text');
  test(() => {
  assert_selection(
    '<div contenteditable><blockquote>|<img src="../resources/abe.png"></blockquote></div>',
    'indent',
    '<div contenteditable><blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;"><blockquote>|<img src="../resources/abe.png"></blockquote></blockquote></div>');
}, 'indent <blockquote> with an image');
</script>
